Add CommonMessages, refactor HmdaFileUpload accordingly
