+++ api/src/main/scala/hmda/api/processing/submission/CommonMessages.scala
@@ -1,0 +1,8 @@
+package hmda.api.processing.submission
+
+object CommonMessages {
+  trait Command
+  trait Event
+  case object Shutdown
+  case object GetState
+}
